Allow using custom slug for Dashboard Views
By defalt, the slug will be a URL-safe name of the
Admin menu as configured under GravityView's global settings;
this slug can also be changed using a filter.

This also adds helper methods to get the submenu slug, as well
as to check if the submenu is a Dashboard View.